backend/requirements.txt
asgiref==3.6.0
astroid==2.15.5
autopep8==2.0.2
bcrypt==4.0.1
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
connection==2021.7.20
DateTime==5.1
dill==0.3.6
Django==4.0.2
django-cors-headers==3.14.0
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
path==16.6.0
platformdirs==3.5.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pylint==2.17.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
zope.interface==6.0
